fixed mass issues
Did a lot of testing with ART on getting the units correct. that should be now taken care of. All tests pass right now.

The test_snia_rate_code_units was removed because it was wrong to begin with. I needed to mess with code mass, not code time. I changed how the ART code works to make that cleaner and this was unncessary.