updates tests for stdout enrichment, includes SN II
This is an attempt to get this right, I didn't actually test it, since
I'm on shangrila at the moment and don't want to set up the environment.

This fixes typos in the AGB tests from the last commit, and adds the
appropriate tests for SN II.

The tests against my Python code are not going to work. Those
defintiely need to be fixed.