stdout checks for feedback
This took many changes. The biggest one was adding a new unit set to handle the energies. I added tests for all the elements of interest. some tests changed since I'm now doing everything in code units in ART, so the conversions are less easy to check now.

The stdout file is from the first stars in the sim, so they don't pass the criteria that we need AGB stars. Those tests pass only because everything is zero. A future update will provide a new stdout file.